Add sealed trait and object UUIDOrIntOrLongOrString
For handling ids of type other than UUID, with UUID as default.

Change in TableSettings.scala.

Handle incoming types in DatabaseTablesDAO.scala

But! After writing these changes I realized this data comes in as a
string ALWAYS from the URL parameter. So I will scrap all this and see
if all that is needed is a validator to convert the string to its
correct type. Maybe nothing needs to be done at all except remove the
UUID.fromString call from the prepared statement near
DatabaseTablesDAO.scala:160